refactor: Move semantic validations to dedicated file and reduce complexity
Moves the five semantic validation methods out of crossref_validator.go
(which exceeded the 800-line service convention limit) into a new
semantic_validator.go file.

Refactors validateValuationRuleCycles to reduce cognitive complexity below
the gocognit threshold (was 30, limit is 20) by extracting:
- buildValuationRuleGraph: constructs the adjacency list
- valuationRuleDFS: iterative DFS traversal
- reportValuationCycle: records and deduplicates cycle errors

No behaviour change; all existing tests continue to pass.
